Add review to recipes cooked and retrieve data

+++ app/src/main/java/ar/uba/fi/remy/model/ReviewItem.kt
@@ -1,4 +1,4 @@
 package ar.uba.fi.remy.model
 
-class ReviewItem(var id: Int, var nombre:String, var puntaje:Int, var foto: String) {
+class ReviewItem(var id: Int, var nombre:String, var puntaje: Float, var foto: String) {
 }
